requirements.txt
Scrapy>=2.11,<3.0
scrapy-impersonate>=1.0,<2.0
scrapy-playwright>=0.0.40,<1.0
playwright>=1.40,<2.0
curl_cffi>=0.7,<1.0
pymongo>=4.6,<5.0
minio>=7.2,<8.0
dagster>=1.9,<2.0
dagster-webserver>=1.9,<2.0
beautifulsoup4>=4.12,<5.0
lxml>=5.0,<7.0
html5lib>=1.1,<2.0
readability-lxml>=0.8,<1.0
python-dotenv>=1.0,<2.0
python-json-logger>=3.0,<5.0
tenacity>=8.0,<10.0
requests>=2.31,<3.0
twisted>=24.0,<26.0
python-dateutil>=2.8,<3.0
polars>=1.0,<2.0